requirements-dev.txt
# ============= requirements.txt =============
# Core Streamlit
streamlit>=1.45.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.3.0
langgraph>=0.6.8
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0

# LLM providers
openai>=1.0.0
ollama>=0.6.0

# Vector store
qdrant-client>=1.15.0

# Embeddings
fastembed>=0.2.0

# Authentication (NEW - REQUIRED)
streamlit-authenticator>=0.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
typing-extensions>=4.0.0